third_party/WebKit/LayoutTests/external/wpt/css/css-backgrounds/parsing/border-radius-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['border-radius'] = "1px" should set the property value
PASS e.style['border-radius'] = "1px 2% 3px 4%" should set the property value
FAIL e.style['border-radius'] = "5em / 1px 2% 3px 4%" should set the property value assert_equals: serialization should be canonical expected "5em / 1px 2% 3px 4%" but got "5em 1px 5em 2% 5em 3px 5em 4%"
FAIL e.style['border-radius'] = "1px 2% 3px 4% / 5em" should set the property value assert_equals: serialization should be canonical expected "1px 2% 3px 4% / 5em" but got "1px 5em 2% 5em 3px 5em 4% 5em"
PASS e.style['border-top-left-radius'] = "10px" should set the property value
PASS e.style['border-top-right-radius'] = "20%" should set the property value
PASS e.style['border-bottom-right-radius'] = "30px 40%" should set the property value
PASS e.style['border-bottom-left-radius'] = "50% 60px" should set the property value
Harness: the test ran to completion.
